requirements.txt
paddlepaddle>=2.5.0
paddlenlp>=2.6.0
scikit-learn>=1.0.0
numpy>=1.20.0
